hotfix: default locale
